+++ CHANGELOG.md
@@ -4,14 +4,22 @@ NEW FEATURES:
 
 * Add CI tests against go1.10. Drop support for go1.8. ([#1620](https://github.com/golang/dep/pull/1620))
 * Added `install.sh` script. ([#1533](https://github.com/golang/dep/pull/1533))
+* List out of date projects in dep status. ([#1553](https://github.com/golang/dep/pull/1553)).
 * Enabled opt-in persistent caching via $DEPCACHEAGE env var. ([#1711](https://github.com/golang/dep/pull/1711))
 
 BUG FIXES:
 
 IMPROVEMENTS:
 
+* Add template operations support in dep status template output.
+([#1549](https://github.com/golang/dep/pull/1549)).
 * Reduce network access by trusting local source information and only pulling
 from upstream when necessary ([#1250](https://github.com/golang/dep/pull/1250)).
+* Update our dependency on Masterminds/semver to follow upstream again now that [Masterminds/semver#67](https://github.com/Masterminds/semver/pull/67) is merged([#1792](https://github.com/golang/dep/pull/1792)).
+
+WIP:
+* Enable importing external configuration from dependencies during init (#1277). This
+is feature flagged and disabled by default.
 
 # v0.4.1